Compute mesh triangles plane equations

+++ lib/src/utils/OBJ.hpp
@@ -11,7 +11,7 @@ class OBJ
 
 public:
 
-    OBJ(std::vector<glm::vec3> &_triangles, std::vector<glm::vec3> &_vertexes, std::vector<glm::vec2> &_uvs, std::vector<glm::vec3> &_normals, std::vector<unsigned int> &_indexes, std::vector<Material> &_materials)
+    OBJ(std::vector<glm::uvec3> &_triangles, std::vector<glm::vec3> &_vertexes, std::vector<glm::vec2> &_uvs, std::vector<glm::vec3> &_normals, std::vector<unsigned int> &_indexes, std::vector<Material> &_materials)
         : triangles(_triangles)
         , vertexes(_vertexes)
         , uvs(_uvs)
@@ -39,7 +39,7 @@ private:
     void addPoint(unsigned int values[3]);
     void addTriangle(unsigned int values[3]);
 
-    std::vector<glm::vec3> &triangles;
+    std::vector<glm::uvec3> &triangles;
     std::vector<glm::vec3> &vertexes;
     std::vector<glm::vec2> &uvs;
     std::vector<glm::vec3> &normals;
